move functions from code module to guppy_wrapper
